update from origin

+++ fannie/cron/nightly.specialorder.php
@@ -31,21 +31,27 @@
 
 */
 
-include('../config.php');
-include($FANNIE_ROOT.'src/SQLManager.php');
-include($FANNIE_ROOT.'src/cron_msg.php');
+include(dirname(__FILE__) . '/../config.php');
+if (!class_exists('FannieAPI')) {
+    include($FANNIE_ROOT . 'classlib2.0/FannieAPI.php');
+}
+if (!function_exists('cron_msg')) {
+    include($FANNIE_ROOT.'src/cron_msg.php');
+}
 
 set_time_limit(0);
 
 // clean cache
 $cachepath = sys_get_temp_dir()."/ordercache/";
-$dh = opendir($cachepath);
-while (($file = readdir($dh)) !== false) {
-    if ($file == "." || $file == "..") continue;
-    if (!is_file($cachepath.$file)) continue;
-    unlink($cachepath.$file);
+if (file_exists($cachepath)) {
+    $dh = opendir($cachepath);
+    while (($file = readdir($dh)) !== false) {
+        if ($file == "." || $file == "..") continue;
+        if (!is_file($cachepath.$file)) continue;
+        unlink($cachepath.$file);
+    }
+    closedir($dh);
 }
-closedir($dh);
 
 $sql = new SQLManager($FANNIE_SERVER,$FANNIE_SERVER_DBMS,$FANNIE_TRANS_DB,
         $FANNIE_SERVER_USER,$FANNIE_SERVER_PW);
